scene-tree: Optimize rendering of SPBs
If a wayland client wants to draw a solid-color rectangle it has to do
this by submitting a buffer.  The Single Pixel Buffer (SPB) protocol
provides a neat way for it to create 1x1 buffers for this purpose.  This
patch improves how these buffers are then handled in the scene-tree
renderer.

We already ignore opaque black rectangles at the very bottom (and
anything under them) because we assume we'll be rendering on a black
background.  This patch detects black opaque SPBs and handles them in
the same way as black opaque rectangles.  It also detects SPBs and
renders them as rectangles rather than buffers because this may be more
efficient in the underlying renderer.

With the gles2 renderer the original buffer can be released and
destroyed after the texture has been uploaded so we won't be able to
access the wlr_single_pixel_buffer_v1 later on.  I cache the SPB status
and color in the wlr_texture so that we can still access it after the
original SPB is gone.